pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "notion-to-hexo"
version = "3.0.0"
description = "Convert Notion pages to Hexo blog posts with automatic image upload to Aliyun OSS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "oss2>=2.15.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = ["dashscope>=1.14.0"]
ui = ["streamlit>=1.30.0"]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "responses>=0.23",
]

[project.scripts]
notion-to-hexo = "notion_to_hexo.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
